Second batch of zot
Define the cart predicate along with the uniqueness property.
Define the cart-moved predicate.
Define the grid and the adjacent predicate.

Create the zot/clean-history.py script in order to clean the history
from all the adjacent predicates that are repeated at each time instant.

To run the zot checker:

```
./run.sh zot test.lisp
python zot/clean-history.py
```