Fix the zot model
The predicates for cart, arm, cart-speed and arm-speed have
been changed to be temporal variables. To aid in the conversion
new functions have been defined in place of the direct usage of
the change form. This should allow easier switching in case it
will be needed in the future.

The arm movement is now related to both cart movement and the operator
touching the arm. Since arm speed, cart speed and touching predicate are
mutually exclusive the movement can be modeled by a simple or of the
possibities.